requirements.txt
langchain>=0.1.0
youtube-transcript-api>=0.6.1
chromadb>=0.4.22
python-dotenv>=1.0.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
langchain-community>=0.0.10
langdetect>=1.0.9 